requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pytesseract==0.3.10
Pillow==11.0.0
pdfplumber==0.11.0
